refactor(injection): split copying schema and term support
In order to provide a realistic raw schema for user access in the
future it became important to split the concept of copying off that
schema and injecting term support into the schema (only of use
internal to this project). The first method is a bit of a misnomer
in that an optional `id` is injected into the copied schema if
one is not provided, but this should be accounted for by the user
anyway since rethink will automatically create an `id` field if no
primary key is defined.